forgot a cast

+++ examples/ldnsd.c
@@ -220,7 +220,7 @@ main(int argc, char **argv)
 		if (status != LDNS_STATUS_OK) {
 			printf("Error creating answer: %s\n", ldns_get_errorstr_by_id(status));
 		} else {
-			nb = (size_t) sendto(sock, outbuf, answer_size, 0, &addr_him, hislen);
+			nb = sendto(sock, outbuf, answer_size, 0, &addr_him, hislen);
 		}
 		
 		ldns_pkt_free(query_pkt);
